+++ src/main/java/com/multicampus/topicsation/controller/MemberManageController.java
@@ -1,7 +1,6 @@
 package com.multicampus.topicsation.controller;
 
 import com.multicampus.topicsation.dto.SignUpDTO;
-import com.multicampus.topicsation.service.ISignUpService;
 import com.multicampus.topicsation.dto.LoginDTO;
 import com.multicampus.topicsation.dto.MailDTO;
 import com.multicampus.topicsation.service.IMemberManageService;
@@ -22,9 +21,6 @@ import java.util.Map;
 @Controller
 @RequestMapping("/members")
 public class MemberManageController {
-
-    @Autowired
-    private ISignUpService signUpService;
 
     @Autowired
     private IMemberManageService memberManageservice;
@@ -154,7 +150,7 @@ public class MemberManageController {
 
         @PostMapping("/signup-tutees.post")
         public String signUpTutee(@RequestBody SignUpDTO signUpDTO) {
-            boolean result = signUpService.signUpProcess(signUpDTO);
+            boolean result = memberManageservice.signUpProcess(signUpDTO);
             if (result) {
                 return signUpDTO.getEmail();
             } else {
@@ -188,7 +184,7 @@ public class MemberManageController {
 
             System.out.println(signUpDTO);
 
-            boolean result = signUpService.signUpProcess(signUpDTO);
+            boolean result = memberManageservice.signUpProcess(signUpDTO);
             if (result) {
                 return new ResponseEntity <String> (signUpDTO.getEmail(), HttpStatus.OK);
             } else {
@@ -198,7 +194,7 @@ public class MemberManageController {
 
         @PostMapping("/email.send")
         public String emailAuth(@RequestBody MailDTO mailDTO) {
-            boolean result = signUpService.sendMail(mailDTO);
+            boolean result = memberManageservice.signupSendMail(mailDTO);
             if(result){
                 return mailDTO.getAuthKey();
             }
@@ -209,7 +205,7 @@ public class MemberManageController {
 
         @PostMapping("/signup/success.post")
         public String successEmailAuth(@RequestBody SignUpDTO signUpDTO) {
-            boolean result = signUpService.successEmailAuth(signUpDTO);
+            boolean result = memberManageservice.isSuccessEmailAuth(signUpDTO);
             if(result){
                 return "emailAuthSuccess";
             } else {
