+++ src/main/resources/mybatis/mapper/SignUpDAO.xml
@@ -31,10 +31,10 @@
         insert into TUTOR_INFO (tutor_id, gender, nationality, certificate) values (last_insert_id(), #{gender}, #{nationality}, #{certificate})
     </insert>
 <!--    회원가입시 이메일인증 여부 갱신 -->
-    <update id="successEmailAuthDAO" parameterType="String">
+    <update id="isSuccessEmailAuthDAO" parameterType="String">
         update USER set email_auth = 1 where BINARY email = #{email}
     </update>
-
+<!--    튜터 인증파일 파일명을 user_id로 변경하기 위해-->
     <select id="getUserId" parameterType="String" resultType="String">
         select user_id FROM USER where BINARY email = #{email}
     </select>
